requirements.txt
yt-dlp
openai-whisper
anthropic
openai
python-dotenv
rich
librosa
numpy
google-genai
demucs
